requirements.txt

pyserial
# pypubsub
